requirements.txt
panel
holoviews
hvplot
streamlit_option_menu
streamlit-aggrid==0.3.3
datetime
pandas_profiling
matplotlib
folium>=0.13
jinja2
branca
streamlit_folium
xlrd
openpyxl
streamlit-extras
streamlit-pandas-profiling
sweetviz